src/iot_stick/requirements.txt
Django==1.11.6
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.2
requests==2.18.4
urllib3==1.22
django-filter==1.1.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
drf-writable-nested==0.3.1
markdown==2.6.9
celery==4.1.0
rabbitmq==0.2.0
uwsgi==2.0.15

